+++ src/get_fignums_example.cpp
@@ -7,19 +7,20 @@
 #include <print>
 
 using matplotlibcpp::Figure;
+using Vec = std::vector<double>;
 
 int main() {
 
     matplotlibcpp::PyPlot plt;
 
     Figure fig1 = plt.figure({.num = 1, .label = "sine_plot"});
-    plt.plot({{1, 2, 3}, {1, 4, 9}});
+    plt.plot({.x = Vec{1, 2, 3}, .y = Vec{1, 4, 9}});
 
     Figure fig2 = plt.figure({.num = 2, .label = "cosine_plot"});
-    plt.plot({{1, 2, 3}, {1, 2, 3}});
+    plt.plot({.x = Vec{1, 2, 3}, .y = Vec{1, 2, 3}});
 
     auto fig3 = plt.figure({.num = 3, .label = "bar_plot"});
-    plt.plot({{1, 2, 3}, {5, 3, 7}});
+    plt.plot({.x = Vec{1, 2, 3}, .y = Vec{5, 3, 7}});
 
 
     std::vector<int> nums = plt.get_fignums();